requirements.txt
python-vlc >=3.0
pytchat >=0.5.5